chore: lint codebase
We're currently on a drive to lint our code more, so I decided to begin by
autofixing what I could using `rubocop --autofix`.

I also have the VSCode rubocop extension active, which offers further
non-essential suggestions, like prefering single quotation marks.

Main thing to say here is:

  A. It would be very surprising for any impactful change to be present here,
  since these are all autofix changes, and
  B. This is not an exhaustive linting of the codebase, because I only
  corrected the VSCode rubocop extension suggestions for the files I
  happened to be in.

So at this point in time, it's on our to-do list to completely lint these
rest of the codebase. But this is a good start.